+++ examples/demo-rollup/src/lib.rs
@@ -1,24 +1,16 @@
 #![deny(missing_docs)]
 #![doc = include_str!("../README.md")]
 
-use std::env;
-use std::str::FromStr;
-
 use const_rollup_config::{ROLLUP_BATCH_NAMESPACE_RAW, ROLLUP_PROOF_NAMESPACE_RAW};
 use sov_celestia_adapter::types::Namespace;
 
 mod mock_rollup;
 
 pub use mock_rollup::*;
-use tracing::info;
-use tracing_subscriber::layer::SubscriberExt;
-use tracing_subscriber::util::SubscriberInitExt;
-use tracing_subscriber::{fmt, EnvFilter};
 
 mod celestia_rollup;
 
 pub use celestia_rollup::*;
-use sov_modules_rollup_blueprint::DEFAULT_SOV_ROLLUP_LOGGING;
 
 mod eth;
 
@@ -28,29 +20,3 @@ pub const ROLLUP_BATCH_NAMESPACE: Namespace = Namespace::const_v0(ROLLUP_BATCH_N
 
 /// The rollup stores the zk proofs in the namespace b"sov-test-p" on Celestia.
 pub const ROLLUP_PROOF_NAMESPACE: Namespace = Namespace::const_v0(ROLLUP_PROOF_NAMESPACE_RAW);
-
-/// Default initialization of logging
-pub fn initialize_logging() {
-    let tracing_layers = tracing_subscriber::registry().with(fmt::layer()).with(
-        EnvFilter::from_str(
-            &env::var("RUST_LOG").unwrap_or_else(|_| DEFAULT_SOV_ROLLUP_LOGGING.to_string()),
-        )
-        .unwrap(),
-    );
-
-    if cfg!(tokio_unstable) {
-        tracing_layers.with(console_subscriber::spawn()).init();
-    } else {
-        tracing_layers.init();
-        info!(
-            info_url = "https://github.com/tokio-rs/console",
-            "The Tokio debugging console will not be available; must compile with `cfg(tokio_unstable)` to enable"
-        );
-    }
-
-    let prev_hook = std::panic::take_hook();
-    std::panic::set_hook(Box::new(move |panic_info| {
-        tracing_panic::panic_hook(panic_info);
-        prev_hook(panic_info);
-    }));
-}